==> 3.1_comorbidity_data.ipynb <==
# Libraries
import numpy as np
import pandas as pd
pd.set_option('display.max_columns', 40)
pd.set_option('display.width', 2000)
import torch
import pickle

from IPython.core.interactiveshell import InteractiveShell
InteractiveShell.ast_node_interactivity = "all"

device = torch.device("cuda" if torch.cuda.is_available() else "cpu")

# Import
path = r'switch_data/final_trimmed_snomed_embedding_128d-copy.csv'
embedding = pd.read_csv(path)

# Import
path = r'switch_data/final_problem_dummies.csv'
problem_dummies = pd.read_csv(path)

# Import
path = r'switch_data/icare_switch_data.csv'
icare_df = pd.read_csv(path)
# Import
path = r'switch_data/icare_switch_data_preprocessed_2023.csv'
icare_df_2023 = pd.read_csv(path)
# Concat
joint_switch = pd.concat([icare_df, icare_df_2023])

import json, snowflake.connector

# establish the connection to snowflake
ctx = snowflake.connector.connect( 
    **json.load(open('/opt/ich/python-snowflake-defaults.json')))
    
# verify and test if connection is working
try: 
    cs = ctx.cursor() 
    cs.execute('SELECT current_version(), current_role(), current_warehouse()')
    print(cs.fetchone())
finally: 
    cs.close()

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.EPISODES
'''
cur = ctx.cursor().execute(query)
episodes = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.EPISODES_2023
'''
cur = ctx.cursor().execute(query)
episodes_2023 = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Create lists 
subject_list = problem_dummies['SUBJECT'].unique().tolist()
spell_list = joint_switch['SPELL_IDENTIFIER'].unique().tolist()
# Filter 
episodes = episodes[episodes['SUBJECT'].isin(subject_list)]
episodes = episodes[episodes['SPELL_IDENTIFIER'].isin(spell_list)]
episodes_2023 = episodes_2023[episodes_2023['SUBJECT'].isin(subject_list)]
episodes_2023 = episodes_2023[episodes_2023['SPELL_IDENTIFIER'].isin(spell_list)]
# Drop duplicates
episodes.drop_duplicates(subset=['SUBJECT', 'SPELL_IDENTIFIER'], inplace=True)
episodes_2023.drop_duplicates(subset=['SUBJECT', 'SPELL_IDENTIFIER'], inplace=True)

len(subject_list)
len(spell_list)
episodes.SUBJECT.nunique()
episodes.SPELL_IDENTIFIER.nunique()
episodes_2023.SUBJECT.nunique()
episodes_2023.SPELL_IDENTIFIER.nunique()

len(set.intersection(set(episodes['SUBJECT']), set(episodes_2023['SUBJECT'])))
len(set.intersection(set(episodes['SPELL_IDENTIFIER']), set(episodes_2023['SPELL_IDENTIFIER'])))

# Now have unique spell to subject mapping

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.EPISODES_DIAGNOSIS
'''
cur = ctx.cursor().execute(query)
diagnosis = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Filter
diagnosis = diagnosis[diagnosis['SUBJECT'].isin(subject_list)]
diagnosis = diagnosis[diagnosis['SPELL_IDENTIFIER'].isin(spell_list)]

diagnosis.SUBJECT.nunique()
diagnosis.SPELL_IDENTIFIER.nunique()

# Filter by uti
uti_diagnoses = diagnosis[(diagnosis['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("urinary tract infection", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("pyelonephritis", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("urinary tract infection", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("pyelonephritis", case=False))]
# Filter by pneumonia
pneumonia_diagnoses = diagnosis[(diagnosis['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("pneumonia", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("pneumonia", case=False))]

# Filter by sepsis
sepsis_diagnoses = diagnosis[(diagnosis['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("sepsis", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("sepsis", case=False))]

# Filter by cellulitis
cellulitis_diagnoses = diagnosis[(diagnosis['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("cellulitis", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("cellulitis", case=False))]

# Filter by bacteraemia
bacteraemia_diagnoses = diagnosis[(diagnosis['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("blood stream infection", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("bacteraemia", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("blood stream infection", case=False)) | \
                                  (diagnosis['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("bacteraemia", case=False))]

uti_diagnoses.SUBJECT.nunique()
uti_diagnoses.SPELL_IDENTIFIER.nunique()

pneumonia_diagnoses.SUBJECT.nunique()
pneumonia_diagnoses.SPELL_IDENTIFIER.nunique()

sepsis_diagnoses.SUBJECT.nunique()
sepsis_diagnoses.SPELL_IDENTIFIER.nunique()

cellulitis_diagnoses.SUBJECT.nunique()
cellulitis_diagnoses.SPELL_IDENTIFIER.nunique()

bacteraemia_diagnoses.SUBJECT.nunique()
bacteraemia_diagnoses.SPELL_IDENTIFIER.nunique()

735 + 1504 + 1504 + 261
len(set.union(set(uti_diagnoses.SPELL_IDENTIFIER.unique()), set(pneumonia_diagnoses.SPELL_IDENTIFIER.unique()), set(sepsis_diagnoses.SPELL_IDENTIFIER.unique()), set(cellulitis_diagnoses.SPELL_IDENTIFIER.unique())))
len(set.intersection(set(uti_diagnoses.SPELL_IDENTIFIER.unique()), set(pneumonia_diagnoses.SPELL_IDENTIFIER.unique()), set(sepsis_diagnoses.SPELL_IDENTIFIER.unique()), set(cellulitis_diagnoses.SPELL_IDENTIFIER.unique())))

# Create binary columns
uti_diagnoses['uti'] = 1
pneumonia_diagnoses['pneumonia'] = 1
sepsis_diagnoses['sepsis'] = 1
cellulitis_diagnoses['cellulitis'] = 1

# Merge DataFrames
merged_df = uti_diagnoses[['SPELL_IDENTIFIER', 'uti']]
merged_df = merged_df.merge(pneumonia_diagnoses[['SPELL_IDENTIFIER', 'pneumonia']], how='outer')
merged_df = merged_df.merge(sepsis_diagnoses[['SPELL_IDENTIFIER', 'sepsis']], how='outer')
merged_df = merged_df.merge(cellulitis_diagnoses[['SPELL_IDENTIFIER', 'cellulitis']], how='outer')

# Fill NaN values with 0
merged_df = merged_df.fillna(0)
disease_df = merged_df.drop_duplicates()
disease_df.reset_index(inplace=True, drop=True)

# Load lists
with open('problem_dummies_subject_list.pickle', 'rb') as handle:
    subject_list = pickle.load(handle)
with open('joint_switch_spell_list.pickle', 'rb') as handle:
    spell_list = pickle.load(handle)

len(subject_list)
len(spell_list)

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.EPISODES_DIAGNOSIS_2023
'''
cur = ctx.cursor().execute(query)
diagnosis_2023 = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])
# Filter
diagnosis_2023 = diagnosis_2023[diagnosis_2023['SUBJECT'].isin(subject_list)]
diagnosis_2023 = diagnosis_2023[diagnosis_2023['SPELL_IDENTIFIER'].isin(spell_list)]
# Filter by uti
uti_diagnoses = diagnosis_2023[(diagnosis_2023['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("urinary tract infection", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("pyelonephritis", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("urinary tract infection", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("pyelonephritis", case=False))]
# Filter by pneumonia
pneumonia_diagnoses = diagnosis_2023[(diagnosis_2023['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("pneumonia", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("pneumonia", case=False))]

# Filter by sepsis
sepsis_diagnoses = diagnosis_2023[(diagnosis_2023['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("sepsis", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("sepsis", case=False))]

# Filter by cellulitis
cellulitis_diagnoses = diagnosis_2023[(diagnosis_2023['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("cellulitis", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("cellulitis", case=False))]

# Filter by bacteraemia
bacteraemia_diagnoses = diagnosis_2023[(diagnosis_2023['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("blood stream infection", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_ICD'].str.contains("bacteraemia", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("blood stream infection", case=False)) | \
                                  (diagnosis_2023['DIAGNOSIS_DESCRIPTION_SNOMED'].str.contains("bacteraemia", case=False))]

uti_diagnoses.SUBJECT.nunique()
uti_diagnoses.SPELL_IDENTIFIER.nunique()

pneumonia_diagnoses.SUBJECT.nunique()
pneumonia_diagnoses.SPELL_IDENTIFIER.nunique()

sepsis_diagnoses.SUBJECT.nunique()
sepsis_diagnoses.SPELL_IDENTIFIER.nunique()

cellulitis_diagnoses.SUBJECT.nunique()
cellulitis_diagnoses.SPELL_IDENTIFIER.nunique()

bacteraemia_diagnoses.SUBJECT.nunique()
bacteraemia_diagnoses.SPELL_IDENTIFIER.nunique()

# Create binary columns
uti_diagnoses['uti'] = 1
pneumonia_diagnoses['pneumonia'] = 1
sepsis_diagnoses['sepsis'] = 1
cellulitis_diagnoses['cellulitis'] = 1

# Merge DataFrames
merged_df = uti_diagnoses[['SPELL_IDENTIFIER', 'uti']]
merged_df = merged_df.merge(pneumonia_diagnoses[['SPELL_IDENTIFIER', 'pneumonia']], how='outer')
merged_df = merged_df.merge(sepsis_diagnoses[['SPELL_IDENTIFIER', 'sepsis']], how='outer')
merged_df = merged_df.merge(cellulitis_diagnoses[['SPELL_IDENTIFIER', 'cellulitis']], how='outer')

# Fill NaN values with 0
merged_df = merged_df.fillna(0)
disease_df_2023 = merged_df.drop_duplicates()
disease_df_2023.reset_index(inplace=True, drop=True)
disease_df_2023

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.DEMOGRAPHIC
'''
cur = ctx.cursor().execute(query)
demographics = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])

# Create episode lists 
episode_subject_list = episodes['SUBJECT'].unique().tolist()

# Filter
demographics = demographics[demographics['SUBJECT'].isin(episode_subject_list)]
# Filter for relevant columns 
demographics = demographics[['SUBJECT', 'GENDER_DESC', 'AGE', 'ETHNICITY_DESC', 'IMDDECIL']]
# Conver to int
demographics['AGE'] = pd.to_numeric(demographics['AGE'])
demographics['IMDDECIL'] = pd.to_numeric(demographics['IMDDECIL'])
# Reset index 
demographics.reset_index(inplace=True, drop=True)

# Filter for those over 18
demographics = demographics[demographics['AGE'] >= 18]

demographics.AGE.max()
demographics.AGE.min()
demographics.IMDDECIL.max()
demographics.IMDDECIL.min()

from sklearn import preprocessing

# Filling missing data, normalize features
min_max_scaler = preprocessing.MinMaxScaler()
data_scaled = min_max_scaler.fit_transform(demographics[['AGE', 'IMDDECIL']])
data_scaled = np.nan_to_num(data_scaled, nan=-1)

def grouped_ethincity_fun(a):
    if "mixed " in a.lower():
        return "Other"
    elif "white" in a.lower():
        return "White"
    elif "black" in a.lower():
        return "Black"
    elif "asian" in a.lower():
        return "Asian"
    elif "other " in a.lower():
        return "Other"
    else:
        return "Other"

# Group race
demographics["grouped_ethincity"] = demographics.ETHNICITY_DESC.apply(lambda x: grouped_ethincity_fun(x))

# Create one hot
one_hot_demographics = pd.get_dummies(demographics[['GENDER_DESC', 'grouped_ethincity']])

# Create final df
final_demographics = pd.DataFrame()
final_demographics['SUBJECT'] = demographics['SUBJECT'] 
final_demographics[['AGE', 'IMDDECIL']] = data_scaled
final_demographics = pd.concat([final_demographics, one_hot_demographics], axis=1)

# Import 
query = '''
SELECT * from ICHT_PROD.ICHT_COVID.DEMOGRAPHIC_2023
'''
cur = ctx.cursor().execute(query)
demographics_2023 = pd.DataFrame.from_records(iter(cur), columns=[x[0] for x in cur.description])
# Create episode lists 
episode_2023_subject_list = episodes_2023['SUBJECT'].unique().tolist()
# Filter
demographics_2023 = demographics_2023[demographics_2023['SUBJECT'].isin(episode_2023_subject_list)]
# Filter for relevant columns 
demographics_2023 = demographics_2023[['SUBJECT', 'GENDER_DESC', 'AGE', 'ETHNICITY_DESC', 'IMDDECIL']]
# Conver to int
demographics_2023['AGE'] = pd.to_numeric(demographics_2023['AGE'])
demographics_2023['IMDDECIL'] = pd.to_numeric(demographics_2023['IMDDECIL'])
# Reset index 
demographics_2023.reset_index(inplace=True, drop=True)
from sklearn import preprocessing
# Filling missing data, normalize features
min_max_scaler = preprocessing.MinMaxScaler()
data_scaled = min_max_scaler.fit_transform(demographics_2023[['AGE', 'IMDDECIL']])
data_scaled = np.nan_to_num(data_scaled, nan=-1)
# Group race
demographics_2023["grouped_ethincity"] = demographics_2023.ETHNICITY_DESC.apply(lambda x: grouped_ethincity_fun(x))
# Create one hot
one_hot_demographics = pd.get_dummies(demographics_2023[['GENDER_DESC', 'grouped_ethincity']])
# Create final df
final_demographics_2023 = pd.DataFrame()
final_demographics_2023['SUBJECT'] = demographics_2023['SUBJECT'] 
final_demographics_2023[['AGE', 'IMDDECIL']] = data_scaled
final_demographics_2023 = pd.concat([final_demographics_2023, one_hot_demographics], axis=1)
final_demographics_2023

demographics_2023.AGE.max()
demographics_2023.AGE.min()
demographics_2023.IMDDECIL.max()
demographics_2023.IMDDECIL.min()

# Save
#with open('problem_dummies_subject_list.pickle', 'wb') as handle:
#    pickle.dump(subject_list, handle)
#with open('joint_switch_spell_list.pickle', 'wb') as handle:
#    pickle.dump(spell_list, handle)

# Save
#episodes.to_csv('switch_data/chronic_switch_episodes.csv', index=False)
# Save
#disease_df.to_csv('switch_data/chronic_switch_disease.csv', index=False)
# Save
#final_demographics.to_csv('switch_data/chronic_switch_demographics.csv', index=False)

# Save
#episodes_2023.to_csv('switch_data/chronic_switch_episodes_2023.csv', index=False)

# Save
#disease_df_2023.to_csv('switch_data/chronic_switch_disease_2023.csv', index=False)

# Save
#final_demographics_2023.to_csv('switch_data/chronic_switch_demographics_2023.csv', index=False)

# Import
path = r'switch_data/chronic_switch_episodes.csv'
episodes = pd.read_csv(path)
# Import
path = r'switch_data/chronic_switch_episodes_2023.csv'
episodes_2023 = pd.read_csv(path)

# Create episode lists 
episode_subject_list = episodes['SUBJECT'].unique().tolist()
episode_spell_list = episodes['SPELL_IDENTIFIER'].unique().tolist()
# Create episode lists 
episode_2023_subject_list = episodes_2023['SUBJECT'].unique().tolist()
episode_2023_spell_list = episodes_2023['SPELL_IDENTIFIER'].unique().tolist()

# Import
path = r'switch_data/icare_switch_data_preprocessed.csv'
icare_df_preprocessed = pd.read_csv(path)
# Import
path = r'switch_data/icare_switch_data_preprocessed_2023.csv'
icare_df_preprocessed_2023 = pd.read_csv(path)

len(icare_df_preprocessed)
icare_df_preprocessed['SPELL_IDENTIFIER'].nunique()
len(icare_df_preprocessed_2023)
icare_df_preprocessed_2023['SPELL_IDENTIFIER'].nunique()
len(problem_dummies)
problem_dummies['SUBJECT'].nunique()

# Filter
icare_df_preprocessed = icare_df_preprocessed[icare_df_preprocessed['SPELL_IDENTIFIER'].isin(episode_spell_list)]
icare_df_preprocessed_2023 = icare_df_preprocessed_2023[icare_df_preprocessed_2023['SPELL_IDENTIFIER'].isin(episode_2023_spell_list)]
filtered_problem_dummies = problem_dummies[problem_dummies['SUBJECT'].isin(episode_subject_list)]

# Save
icare_df_preprocessed.to_csv('switch_data/chronic_switch_icare_df_preprocessed.csv', index=False)
# Save
icare_df_preprocessed_2023.to_csv('switch_data/chronic_switch_icare_df_preprocessed_2023.csv', index=False)
# Save
filtered_problem_dummies.to_csv('switch_data/chronic_switch_problem_dummies.csv', index=False)
